chore: fix compiler warning in ByteUtils (#18567)

+++ platform-sdk/consensus-utility/src/main/java/org/hiero/consensus/utility/ByteUtils.java
@@ -153,7 +153,7 @@ public final class ByteUtils {
                 if (index >= data.length) {
                     break;
                 }
-                result += (data[index] & 0xff) << (8 * (1 - offset));
+                result += (short) ((data[index] & 0xff) << (8 * (1 - offset)));
             }
             return result;
         }
